rename the script
